requirements.txt
mcp
GitPython
httpx
radon
bandit
pydantic-settings
gradio
gradio_client
pydantic
python-dotenv
pytest
